Rename module: responses -> results
