+++ src/mcp_server/engines/top_row_engine.py
@@ -56,7 +56,7 @@ import pandas as pd
 
 from mcp_server import DataToolOutput
 from mcp_server.engines.filters import build_filter_params, apply_filters, build_filter_doc
-from mcp_server.responses import text_result, force_result
+from mcp_server.results import text_result, force_result
 
 
 def load_top_row_dataset(mcp, config, yaml_path):
