+++ src/mcp_server/engines/unique_values_engine.py
@@ -41,7 +41,7 @@ import pandas as pd
 
 from mcp_server import DataToolOutput
 from mcp_server.engines.filters import build_filter_params, apply_filters, build_filter_doc
-from mcp_server.responses import text_result
+from mcp_server.results import text_result
 
 
 def load_unique_values_dataset(mcp, config, yaml_path):
